refact: share isgotype between all modules
